factor our processSampleBulkAdd() to share with LKB MediaCreatePage.tsx
